+++ App/Targets/three-dollar-in-my-pocket/Sources/SceneDelegate.swift
@@ -20,7 +20,7 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         window = UIWindow(frame: windowScene.coordinateSpace.bounds)
         window?.windowScene = windowScene
         window?.backgroundColor = UIColor.init(r: 28, g: 28, b: 28)
-        window?.rootViewController = SigninViewController.instance()
+        window?.rootViewController = SplashVC.instance()
         window?.makeKeyAndVisible()
         
         self.reserveDynamicLinkIfExisted(connectionOptions: connectionOptions)
